Website url Improvements
- Introduced a centralized URL validation function to improve consistency across components.
- Updated AssistantAdminUI to delay tab activation for better user experience.
- Enhanced WebsiteURLInput and AssistantModal to utilize the new validation logic, ensuring URLs are properly validated and sanitized.
